Update patch set 2
Patch Set 2:

Tests are passing OK without this change. For the wiremock dependency, this is a limitation of the project.py script as test dependencies not present in the acceptance-framework are not added to the classpath. It seems it is not an easy fix according to D. Ostrovsky, so I just posted a workaround here:

https://gerrit-review.googlesource.com/#/c/79230/

Patch-set: 2
Label: Verified=0