Update patch set 1
Patch Set 1:

I was expecting the dependencies to be provided by PLUGIN_DEPS and PLUGIN_TEST_DEPS in a similar way as they are when building with Buck. However, I'm getting this errors [1] which seem to show that's not the case (or most probably, I'm lacking something in the build files). When I try to add the httpclient dependency, I'm getting the same errors but this time with a StackOverflowError [2] which makes me think I'm definitively doing something really wrong.

[1] http://pastebin.com/CGDmt8uF
[2] http://pastebin.com/giL9cCPZ

Any hints about this?

Patch-set: 1