Update patch set 20
Patch Set 20:

I compared the jar generated by buck and bazel and noticed those 2 differences in the jar build by bazel:
-*.java files are in the jar, it this per design?
-manifest contains "Implementation-Vendor: Gerrit Code Review" instead of what is specified in the BUILD file (Implementation-Vendor: Ericsson).

The *.java files made me think that we have a way to generate the source jar in Buck, can we do that in bazel?

The manifest is not a showstopper but we will need to fix this eventually.

Patch-set: 20
Label: Verified=0